+++ themes/default/views/user/account/login.php
@@ -41,6 +41,4 @@
     <?php $this->endWidget(); ?>
 </div><!-- form -->
 
-
-<?php //$this->widget('application.modules.social.extensions.eauth.EAuthWidget',array('popup' => Yii::app()->eauth->popup,'services' => Yii::app()->eauth->getServices()));?>
-
+<?php  $this->widget('application.modules.social.extensions.eauth.EAuthWidget',array('action' => '/social/social/login/'));?>